indentation-c-style: Change get_line_indentation() behaviour.
Change it to not only search on the current line for the first right brace but
search until one is found or a right bracket is found.

This fixes indentation issues with c++11 type return value syntax such as:

auto testFunction()
  -> int {
  return 5;
}

Where the return type is put on another line different from the end right brace.

https://bugzilla.gnome.org/show_bug.cgi?id=722720