Fix Glib-CRITICAL **: Source ID was not found when attempting to remove it
In the symbol db plugin, the callback on_editor_buffer_symbols_update_timeout
was returning FALSE in some case which automatically remove the source.

In this case, we have to clear sdb_plugin->buf_update_timeout_id to indicate
that the source is removed.

In most of the case, I think it was better to return TRUE to keep checking the
buffer even if an update shouldn't be done at the moment.