build: Don't require exact match in the bison check
Autoconf changed its way to use bison several years ago:
https://git.savannah.gnu.org/cgit/autoconf.git/commit/?id=42761668c0300afa7f8bf5ba736458a818cd5d15

It haven't been made available in any release, but it is possible for
distributions to backport the commit into their packages, causing the
bison check to fail. Specifically, it already happens on FreeBSD.

Therefore, instead of requiring an exact match, the check is changed to
only require the variable is started with 'bison' and a white space.